// src/APIs/types/boards/boards.enum.ts
export enum GetBoardsActionTypes {
  REQUESTED_BOARD_GET = "REQUESTED_BOARD_GET",
  RECIVED_BOARD_GET = "RECIVED_BOARD_GET",
  FAILED_BOARD_GET = "FAILED_BOARD_GET",
}

export enum CreateBoardActionTypes {
  REQUESTED_BOARD_CREATE = "REQUESTED_BOARD_CREATE",
  RECIVED_BOARD_CREATE = "RECIVED_BOARD_CREATE",
  FAILED_BOARD_CREATE = "FAILED_BOARD_CREATE",
}

export enum UpdateBoardActionTypes {
  REQUESTED_BOARD_UPDATE = "REQUESTED_BOARD_UPDATE",
  RECIVED_BOARD_UPDATE = "RECIVED_BOARD_UPDATE",
  FAILED_BOARD_UPDATE = "FAILED_BOARD_UPDATE",
}

export enum DeleteBoardAcrionTypes {
  REQUESTED_BOARD_DELETE = "REQUESTED_BOARD_DELETE",
  RECIVED_BOARD_DELETE = "RECIVED_BOARD_DELETE",
  FAILED_BOARD_DELETE = "FAILED_BOARD_DELETE",
}

export enum BoardExistsActionTypes {
  REQUESTED_BOARD_EXISTS = "REQUESTED_BOARD_EXISTS",
  RECIVED_BOARD_EXISTS = "RECIVED_BOARD_EXISTS",
  FAILED_BOARD_EXISTS = "FAILED_BOARD_EXISTS",
}